+++ R/time-zones.r
@@ -28,7 +28,6 @@ with_tz <- function (time, tzone = "") {
     }
     time
   } else {
-    check_tz(tzone)
     if (is.POSIXlt(time)) new <- as.POSIXct(time)
     else new <- time
     attr(new, "tzone") <- tzone
@@ -38,16 +37,17 @@ with_tz <- function (time, tzone = "") {
 
 #' Replace time zone to create new date-time
 #'
-#' force_tz returns a the date-time that has the same clock time as x in the new
-#'  time zone.
-#' Although the new date-time has the same clock time (e.g. the
-#' same values in the year, month, days, etc. elements) it is a
-#' different moment of time than the input date-time. force_tz defaults to the
-#' Universal Coordinated time zone (UTC) when an unrecognized time zone is
-#' inputted. See [Sys.timezone()] for more information on how R
-#' recognizes time zones.
+#' `force_tz` returns the date-time that has the same clock time as input time,
+#'  but in the new time zone. `force_tzs` is the parallel version of `force_tz`
+#'  in a sense that every element from `time` argument is matched with the
+#'  corresponding time zone in `tzones` argument.
 #'
-#' @export force_tz
+#'  Although the new date-time has the same clock time (e.g. the same values in
+#'  the year, month, days, etc. elements) it is a different moment of time than
+#'  the input date-time.
+#'
+#'  As R date-time vectors cannot hold elements with non-uniform time zones,
+#'  `force_tzs` returns a vector with time zone `tzone_out`, UTC by default.
 #'
 #' @param time a POSIXct, POSIXlt, Date, chron date-time object, or a data.frame
 #'   object. When a data.frame all POSIXt elements of a data.frame are processed
@@ -64,10 +64,13 @@ with_tz <- function (time, tzone = "") {
 #' x <- ymd_hms("2009-08-07 00:00:01", tz = "America/New_York")
 #' force_tz(x, "UTC")
 #' force_tz(x, "Europe/Amsterdam")
+#'
 #' ## DST skip:
+#'
 #' y <- ymd_hms("2010-03-14 02:05:05 UTC")
 #' force_tz(y, "America/New_York", roll=FALSE)
 #' force_tz(y, "America/New_York", roll=TRUE)
+#' @export
 force_tz <- function(time, tzone = "", roll = FALSE) {
   if (is.data.frame(time)) {
     for (nm in names(time)) {
@@ -77,13 +80,27 @@ force_tz <- function(time, tzone = "", roll = FALSE) {
     }
     time
   } else {
-    check_tz(tzone)
     out <- C_force_tz(as.POSIXct(time), tz = as.character(tzone), roll)
     reclass_date(out, time)
-    ## update(time, tz = tzone)
   }
 }
 
-check_tz <- function(tz) {}
-
-# Note: alternative method? as.POSIXlt(format(as.POSIXct(x)), tz = tzone)
+#' @param tzones character vector of timezones to be "enforced" on `time` time
+#'   stamps. If `time` and `tzones` lengths differ, usuall R recycling rules
+#'   apply.
+#' @param tzone_out timezone of the returned date-time vector
+#' @rdname force_tz
+#' @examples
+#'
+#' ## Heterogeneous time-zones:
+#'
+#' x <- ymd_hms(c("2009-08-07 00:00:01", "2009-08-07 00:00:01"))
+#' force_tzs(x, tzones = c("America/New_York", "Europe/Amsterdam"))
+#' force_tzs(x, tzones = c("America/New_York", "Europe/Amsterdam"), tzone_out = "America/New_York")
+#' @export
+force_tzs <- function(time, tzones, tzone_out = "UTC", roll = FALSE) {
+  if (length(time) != length(tzones))
+    tzones <- rep_len(tzones, length(time))
+  out <- C_force_tzs(as.POSIXct(time), tzones, tzone_out, roll)
+  reclass_date(out, time)
+}