+++ src/RcppExports.cpp
@@ -40,6 +40,20 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// C_force_tzs
+Rcpp::newDatetimeVector C_force_tzs(const Rcpp::NumericVector dt, const Rcpp::CharacterVector tzs, const Rcpp::CharacterVector tz_out, const bool roll);
+RcppExport SEXP _lubridate_C_force_tzs(SEXP dtSEXP, SEXP tzsSEXP, SEXP tz_outSEXP, SEXP rollSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Rcpp::NumericVector >::type dt(dtSEXP);
+    Rcpp::traits::input_parameter< const Rcpp::CharacterVector >::type tzs(tzsSEXP);
+    Rcpp::traits::input_parameter< const Rcpp::CharacterVector >::type tz_out(tz_outSEXP);
+    Rcpp::traits::input_parameter< const bool >::type roll(rollSEXP);
+    rcpp_result_gen = Rcpp::wrap(C_force_tzs(dt, tzs, tz_out, roll));
+    return rcpp_result_gen;
+END_RCPP
+}
 
 RcppExport SEXP C_make_d(SEXP, SEXP, SEXP);
 RcppExport SEXP C_parse_dt(SEXP, SEXP, SEXP, SEXP, SEXP);
@@ -49,6 +63,7 @@ RcppExport SEXP C_parse_period(SEXP);
 static const R_CallMethodDef CallEntries[] = {
     {"_lubridate_C_update_dt", (DL_FUNC) &_lubridate_C_update_dt, 12},
     {"_lubridate_C_force_tz", (DL_FUNC) &_lubridate_C_force_tz, 3},
+    {"_lubridate_C_force_tzs", (DL_FUNC) &_lubridate_C_force_tzs, 4},
     {"C_make_d",               (DL_FUNC) &C_make_d,                3},
     {"C_parse_dt",             (DL_FUNC) &C_parse_dt,              5},
     {"C_parse_hms",            (DL_FUNC) &C_parse_hms,             2},