+++ tests/testthat/test-timezones.R
@@ -36,23 +36,34 @@ test_that("with_tz handles data.frames", {
   expect_that(df$y, equals(x_out))
 })
 
+test_that("force_tzs works as expected", {
+
+  expect_equal(force_tzs(x, tzones = c("America/New_York", "Europe/Amsterdam")),
+               ymd_hms("2009-08-07 04:00:01 UTC", "2009-08-06 22:00:01 UTC"))
+
+  expect_equal(force_tzs(x, tzones = c("America/New_York", "Europe/Amsterdam"), tzone_out = "America/New_York"),
+               ymd_hms("2009-08-07 00:00:01 EDT", "2009-08-06 18:00:01 EDT", tz = "America/New_York"))
+
+  ## recycling
+  expect_equal(force_tzs(x, tzones = "America/New_York", tzone_out = "UTC"),
+               ymd_hms("2009-08-07 04:00:01 UTC", "2009-08-07 04:00:01 UTC"))
+
+})
+
 test_that("force_tz works as expected", {
   x <- as.POSIXct("2008-08-03 10:01:59", tz = "America/New_York")
-
   expect_that(force_tz(x, "UTC"),
               equals(as.POSIXct(format(as.POSIXct(x)), tz = "UTC")))
 })
 
 test_that("force_tz handles vectors", {
   x <- as.POSIXct(c("2008-08-03 13:01:59", "2009-08-03 10:01:59"), tz = "America/New_York")
-
   expect_that(force_tz(x, "UTC"),
               equals(as.POSIXct(format(as.POSIXct(x)), tz = "UTC")))
 })
 
 test_that("force_tz handles various date-time classes", {
   x <- as.POSIXct("2008-12-03 13:01:59", tz = "America/New_York")
-
   expect_equal(force_tz(as.POSIXlt(x), "UTC"),
     as.POSIXlt(format(x), tz = "UTC"))
 })